% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ranges.R
\name{check_ranges}
\alias{check_ranges}
\title{Check Value Ranges of Relevant Variables}
\usage{
check_ranges(d)
}
\arguments{
\item{d}{A data frame prepared via the \code{prepare_data} function.}
}
\value{
Prints a message listing any out-of-range values and may
terminate execution if invalid data are detected.
}
\description{
Validates whether values in the dataset fall within expected ranges for
analysis. Used primarily as a safeguard to identify possible data entry
errors.
}
\examples{
\dontrun{
p <- data_paths("data-raw")
data <- prepare_data(p)  # Automatically performs range checks
}
}
\seealso{
\code{\link[=prepare_data]{prepare_data()}} is a wrapper of this function.
}
