% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_predictions.R
\name{compute_predictions}
\alias{compute_predictions}
\title{Compute model predictions}
\usage{
compute_predictions(model, seed = NULL, ...)
}
\arguments{
\item{model}{The model}

\item{seed}{Optional seed for reproducibility}

\item{...}{Unused. There to allow for compatibility with \code{run_scoring_rule_pipeline}.}
}
\value{
Tibble with two columns:
\describe{
\item{epred}{Linear prediction on the logit scale}
\item{ppred}{Posterior prediction on the outcome scale}
}
}
\description{
Given a Bayesian "projpred" or "brms" model, computes linear and full
posterior predicitons.
}
