% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{concords}
\alias{concords}
\title{Table of Pairwise Concordance Measures}
\format{
\subsection{\code{concords}}{

A tibble with 7,744 rows and 20 columns:
\describe{
\item{predictor}{Character; name of the algorithm used as a predictor.}
\item{reference}{Character; name of the algorithm used as a reference.}
\item{Kappa}{Character; Cohen’s kappa with its 95\% confidence interval.}
\item{Accuracy}{Character; accuracy estimate with its 95\% confidence interval.}
\item{Kappa_raw}{Numeric; Cohen’s kappa estimate.}
\item{McnemarPValue}{Numeric; p-value from McNemar’s chi-squared test for
symmetry between rows and columns.}
\item{Accuracy_raw}{Numeric; accuracy estimate.}
\item{NoInformationRate_raw}{Numeric; no-information rate, i.e., the accuracy
if all patients were classified as non-PDD.}
\item{AccuracyPValue}{Numeric; p-value from the Exact Binomial Test comparing
accuracy to the no-information rate.}
\item{Sensitivity}{Numeric; sensitivity estimate.}
\item{Specificity}{Numeric; specificity estimate.}
\item{Pos Pred Value}{Numeric; positive predictive value.}
\item{Neg Pred Value}{Numeric; negative predictive value.}
\item{Precision}{Numeric; precision estimate.}
\item{Recall}{Numeric; recall estimate.}
\item{F1}{Numeric; F1 score.}
\item{Prevalence}{Numeric; PDD prevalence according to the reference algorithm.}
\item{Detection Rate}{Numeric; detection rate.}
\item{Detection Prevalence}{Numeric; detection prevalence.}
\item{Balanced Accuracy}{Numeric; balanced accuracy, i.e.,
(sensitivity + specificity) / 2.}
}
}
}
\source{
Generated by the targets pipeline in \verb{_targets.R}.
}
\usage{
concords
}
\description{
A comprehensive table containing all pairwise comparisons between 68
diagnostic algorithms for Parkinson’s Disease Dementia (PDD) evaluated in
the study. The table serves as the basis for Figures 2, A3, A4, and A5, as
well as for other summary results presented throughout the article.
}
\details{
Each algorithm is used once as a reference and once as a predictor, meaning
that each algorithm pair appears twice. Some metrics, such as Cohen’s kappa,
are symmetrical, whereas others (e.g., accuracy, sensitivity, or specificity)
are not.
}
\seealso{
\itemize{
\item \code{\link[psych:kappa]{psych::cohen.kappa()}} computes and documents Cohen’s kappa and related
metrics.
\item \code{\link[caret:confusionMatrix]{caret::confusionMatrix()}} computes and documents the remaining metrics.
}
}
\keyword{datasets}
