% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate_indexes.R
\name{correlate_indexes}
\alias{correlate_indexes}
\title{Compute correlation between test indexes}
\usage{
correlate_indexes(
  d0,
  v = setNames(c("moca_7", "mmse_7", "moca_cloc", "cloc", "vf_s", "vf_k", "moca_5words",
    "mmse_3words", "moca_cube", "mmse_pent", "moca_abs", "moca_anim"), c("MoCA Sevens",
    "MMSE Sevens", "MoCA Clock", "Clock Drawing", "VF S", "VF K", "MoCA 5 words",
    "MMSE 3 words", "MoCA Cube", "MMSE Pent.", "MoCA Abs.", "MoCA An.")),
  ideal = NULL,
  ...
)
}
\arguments{
\item{d0}{Input data.}

\item{v}{Named character vector of variables to be included.}

\item{ideal}{Optionally a matrix of ideal values. Write \code{NULL} for the default
or \code{NA} if no ideal matrix should be plotted.}

\item{...}{Other variables for \code{\link[=cor]{cor()}}}
}
\value{
A list containing:
\describe{
\item{R_observed}{Correlation matrix of observed data}
\item{R_ideal}{Correlation matrix of the "ideal" pattern}
}
}
\description{
Given raw data, computes selected type of pairwise correlations ("Pearson"
by default) of all selected indexes and plots them. Optionally, the user
can specify an "ideal" correlation matrix of the same variables for
comparison purposes.
}
