% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_paths.R
\name{data_paths}
\alias{data_paths}
\title{Define Raw Data Paths}
\usage{
data_paths(dir = "data-raw")
}
\arguments{
\item{dir}{A character string specifying the input directory where raw data
files reside.}
}
\value{
A named character vector with file paths to all required raw data
files.
}
\description{
Prepares a set of file paths pointing to all required raw data files used in
the analysis. The function enforces naming conventions and validates
compatibility between the contents of the specified directory and the expected
structure. It is intended to be used as input for the \code{prepare_data}
function.
}
\examples{
\dontrun{
p    <- data_paths("data-raw")
data <- prepare_data(p)
}
}
\seealso{
\code{\link[=prepare_data]{prepare_data()}} is the next step in analysis pipeline.
}
