% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_scoring_rule.R
\name{derive_scoring_rule}
\alias{derive_scoring_rule}
\title{Derive a scoring rule for PDD}
\usage{
derive_scoring_rule(pt, coefs, inverse = TRUE, ...)
}
\arguments{
\item{pt}{The threshold on probability scale. E.g., derived from
\code{run_roc}.}

\item{coefs}{Model coefficients as computed by \code{extract_coefficients()}.}

\item{inverse}{Should \code{pt} be inversed from probability to logit scale?}

\item{...}{Unused. There to allow for compatibility with
\code{run_scoring_rule_pipeline.}}
}
\value{
List containing:
\itemize{
\item score coefficients
\item cutoff
\item equation (character)
\item integer scoring rule
}
}
\description{
Given a prediction linear logistic model and its ROC analysis results,
computes a scoring rule for deciding a patient suffers probable PDD given
screening criteria.
}
