% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_varsel.R
\name{describe_varsel}
\alias{describe_varsel}
\title{Extract variable selection summaries}
\usage{
describe_varsel(vs)
}
\arguments{
\item{vs}{A vsel object.}
}
\value{
List containing:
\describe{
\item{n_suggest}{Number of terms suggested for projection.}
\item{plot_pp}{Plot of predictive performance.}
\item{plot_rk}{Plot of predictors' ranking variability.}
\item{plot_rk_cum}{Plot of predictors' cummulative ranking variability.}
}
}
\description{
Given a vsel object, returns basic descriptions of the model to aid
variable selection and projective prediction process.
}
