% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_pdd_sample.R
\name{diagnose_pdd_sample}
\alias{diagnose_pdd_sample}
\title{Diagnose All Patients}
\usage{
diagnose_pdd_sample(d0)
}
\arguments{
\item{d0}{A tibble containing the dataset prepared using \code{prepare_data}.}
}
\value{
A list with two tibbles:
\describe{
\item{\code{algorithms}}{A tibble containing the full specification of all
algorithms used.}
\item{\code{diagnoses}}{A tibble containing:
\itemize{
\item Patient ID
\item Criterion label (linked to the specification)
\item PDD diagnosis (\code{TRUE}/\code{FALSE})
\item Individual diagnostic flags corresponding to Table 2 in Dubois et
al. \href{https://doi.org/10.1002/mds.21844}{(2007)}
\item Impairment labels across cognitive domains (per Level I criteria)
}}
}
}
\description{
Applies all available diagnostic algorithms for probable Parkinson’s disease
dementia (PDD) across the full dataset. The set of criteria is generated
internally via the \code{specify_algorithms} function. Each patient is assessed
using each specification, and the corresponding diagnoses and supporting
information are recorded.
}
\examples{
\dontrun{
p <- data_paths("data-raw")
data <- prepare_data(p)
pdd  <- diagnose_pdd_sample(data)
}

}
\references{
Dubois, B., Burn, D., Goetz, C., Aarsland, D., Brown, R. G., Broe, G. A., ...
& Emre, M. (2007). Diagnostic procedures for Parkinson's disease dementia:
recommendations from the movement disorder society task force. Movement
disorders, 22(16), 2314-2324. \doi{10.1002/mds.21844}
}
\seealso{
\itemize{
\item \code{\link[=prepare_data]{prepare_data()}} prepares \code{d0}.
\item \code{\link[=specify_algorithms]{specify_algorithms()}} runs internally to prepare a set of algorithms to be
used.
\item \code{\link[=diagnose_pdd_case]{diagnose_pdd_case()}} runs internally to diagnose patients using a single
algorithm.
}
}
