% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_incomplete.R
\name{drop_incomplete}
\alias{drop_incomplete}
\title{Drop patients with missing data}
\usage{
drop_incomplete(d)
}
\arguments{
\item{d}{Data set generated by \code{diagnose_pdd_sample}}
}
\value{
The same data with less cases.
}
\description{
Drops patients with any data on PDD missing.
}
