% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_coefficients.R
\name{extract_coefficients}
\alias{extract_coefficients}
\title{Extract point estimate of model coefficients}
\usage{
extract_coefficients(mod, scls, stat = "mean")
}
\arguments{
\item{mod}{Model object or posterior draws compatible with
\code{posterior::as_draws_matrix()}.}

\item{scls}{Scaling values as exported by \code{fit_reference}.}

\item{stat}{Summary statistic applied to posterior draws.}
}
\value{
Named numeric vector of coefficients on the original scale.
}
\description{
Given posterior draws from a {brms} model or {projpred} projection and
scaling values for predictors, converts coefficients from scaled predictors
to the original variable scale and returns summary statistics.
}
