% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_incomplete.R
\name{find_incomplete}
\alias{find_incomplete}
\title{Find patients with missing data}
\usage{
find_incomplete(d)
}
\arguments{
\item{d}{Data set generated by \code{diagnose_pdd_sample}}
}
\value{
Charcter vector with incomplete cases' ids.
}
\description{
Finds patients with any data on PDD missing.
}
