% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_reference.R
\name{fit_reference}
\alias{fit_reference}
\title{Fit Reference Model for Projective Prediction}
\usage{
fit_reference(
  d,
  gs = "Lvl.II (1)",
  y = "PDD",
  X = c("moca_7", "moca_cloc", "vf_k", "moca_5words", "moca_cube", "moca_abs",
    "moca_anim", "mmse_7", "cloc", "vf_s", "mmse_3words", "mmse_pent"),
  expect_nonzero = 5,
  ...
)
}
\arguments{
\item{d}{Data such as those calculated by \code{diagnose_pdd_sample}.}

\item{gs}{Name the "gold standard" variable has in \code{d$type}.}

\item{y}{Name of the PDD diagnosis variable in \code{d}}

\item{X}{Character vector containing all predictor variables from \code{d}.}

\item{expect_nonzero}{Number of variables expected to be truly predictive.}

\item{...}{Other parameters for \code{\link[brms:brm]{brms::brm()}}.}
}
\value{
List with three components:
\describe{
\item{model}{brmsfit with the fitted model}
\item{data}{Tibble with raw input data}
\item{scaling}{Tibble with predictors' mean and SDs}
}
}
\description{
Using a regularised horseshoe prior, fits a full reference model of selected
"gold standard" PDD diagnosis using specified predictor set.
}
\details{
...
}
\seealso{
\itemize{
\item \code{\link[brms:horseshoe]{brms::horseshoe()}} explains the prior implementation.
}
}
