% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_models.R
\name{grid_models}
\alias{grid_models}
\title{Fit a grid of reference models}
\usage{
grid_models(
  d,
  gss = c("Lvl.II (1)", "Lvl.II (2)"),
  N = 5,
  n_chosen = NA,
  bind = FALSE,
  ...
)
}
\arguments{
\item{d}{Input data, e.g., generated by \code{diagnose_pdd_sample}.}

\item{gss}{Gold standard variables.}

\item{N}{Vector of expected non-zero coefficient for a regularised horseshoe
prior.}

\item{n_chosen}{Number of selected parameters.}

\item{bind}{Should all combinations of \code{gss} and \code{N} be found (\code{FALSE}) or
should they be simply binded (\code{TRUE})}

\item{...}{Other parameters for \code{\link[=fit_reference]{fit_reference()}}.}
}
\value{
A list with four components:
\describe{
\item{grid}{A grid of input values.}
\item{reference}{List of reference models.}
\item{vs_object}{List of variable selection results}
\item{vs_plots}{Plots visualising variable selection}
\item{projection}{List of selected models projected onto appropriate
reference model}
}
}
\description{
Using different values of "gold standard" measures of the outcome and number
of expected non-zero parameters, fits a set of reference models via
\code{fit_reference}.
}
\seealso{
\itemize{
\item \code{\link[=fit_reference]{fit_reference()}} for reference model fitting
\item \code{\link[=perform_varsel]{perform_varsel()}} for variable selection implementation
\item \code{\link[projpred:project]{projpred::project()}} for projective prediction implementation
}
}
