% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_apa_table.R
\name{gt_apa_table}
\alias{gt_apa_table}
\title{Generate an APA-style Table}
\usage{
gt_apa_table(x, grp = NULL, nms = NULL, tit = "")
}
\arguments{
\item{x}{A tibble or data frame to be formatted as an APA-style table.}

\item{grp}{An optional column name (as a string or symbol) to use for grouping
rows via \code{gt::gt()}'s \code{groupname_col} parameter. Defaults to
\code{NULL}.}

\item{nms}{An optional column name (as a string or symbol) to use for row
names via \code{gt::gt()}'s \code{rowname_col} parameter. Defaults to
\code{NULL}.}

\item{tit}{A character string specifying the table title. Will be converted
to HTML format. Defaults to \code{""}.}
}
\value{
A \code{gt} table object formatted according to APA style.
}
\description{
Converts a tibble or data frame into an APA-formatted table using the
gt package. The resulting object inherits all the features and
customizations provided by \code{gt}.
}
