% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_item_data.R
\name{import_item_data}
\alias{import_item_data}
\title{Import Item Data}
\usage{
import_item_data(path)
}
\arguments{
\item{path}{A character string specifying the path to the data file.
Typically generated by the \code{data_paths} function.}
}
\value{
A tibble containing item-level data.
}
\description{
Imports item-level data from a specified file path, checks for unrealistic
values, and prepares the data for further processing. This function is
intended to be called within the larger \code{prepare_data} workflow.
}
\examples{
\dontrun{
p <- data_paths("data-raw")
data <- import_item_data(p[1])
}

}
\seealso{
\code{\link[=prepare_data]{prepare_data()}} is a wrapper of this function.
}
