% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_algorithms.r
\name{list_algorithms}
\alias{list_algorithms}
\title{Prepare a List of Algorithms}
\usage{
list_algorithms(rates)
}
\arguments{
\item{rates}{A tibble containing raw summaries of PDD rates as
generated by \code{summarise_rates}.}
}
\value{
A list with two components:
\describe{
\item{\code{faq_tot}}{A character vector containing names of
algorithms that defined IADL deficit by FAQ total score > 7.}
\item{\code{faq_9}}{A character vector containing names of
algorithms that defined IADL deficit by FAQ item 9 > 1.}
}
}
\description{
Using the table with PDD rates  estimates per diagnostic algorithm
generated by \code{summarise_rates}, this function extracts a list of
algorithms using different IADL criteria for PDD. If there was an algorithm
that uses both criteria at the same time, the function prints an error.
}
\examples{
\dontrun{
p <- data_paths("data-raw")
data <- prepare_data(p)
pdd  <- diagnose_pdd_sample(data)
vars <- here::here("data-raw", "VariablesOfInterest.csv")

rates <- summarise_rates(pdd, vars)
algos <- list_algorithms(rates$table)
}

}
\seealso{
\code{\link[=summarise_rates]{summarise_rates()}} which is fed into this function.
}
