% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dendrogram.r
\name{make_dendrogram}
\alias{make_dendrogram}
\title{Make a Dendrogram of PDD Algorithms}
\usage{
make_dendrogram(tab, top)
}
\arguments{
\item{tab}{A table containing all criteria used as generated by
\code{summarise_rates}.}

\item{top}{A tibble containing top criteria for predicting PDD
diagnosis by Level II a generated by \code{table_levelII_approximations}.}
}
\value{
A ggplot object with a dendrogram created
by {ggraph}.
}
\description{
The functions prepares a dendrogram (tree plot) that summarises PDD
algorithms construction starting from IADL operationalisation at the top
followed successively by operationalisation of Global cognitive performance,
Attention, Executive Function, Construction, Memory and Language. The
dendrogram visualises PDD rate estimated by each algorithm via point size.
}
\examples{
\dontrun{
# Raw data:
p <- data_paths("data-raw")
data <- prepare_data(p)
pdd  <- diagnose_pdd_sample(data)

# PDD rates for point size:
vars <- here::here("data-raw", "VariablesOfInterest.csv")
rates <- summarise_rates(pdd, vars)

# Concordance with Level II for edge width:
algos <- list_algorithms(rates$table)
data(concords)
best <- table_levelII_approximations(concords, algos, "Accuracy_raw")

tree <- make_dendrogram(rates$table, best$table)
}

}
\seealso{
\itemize{
\item \code{\link[=summarise_rates]{summarise_rates()}} prepares \code{tab}.
\item \code{\link[=table_levelII_approximations]{table_levelII_approximations()}} prepares \code{top}.
}
}
