% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_projpred.R
\name{perform_projpred}
\alias{perform_projpred}
\title{Perform projective prediction}
\usage{
perform_projpred(vs, n_chosen = NULL, seed = NA, ...)
}
\arguments{
\item{vs}{A vsel object.}

\item{n_chosen}{Number of parameters to retain in the model.}

\item{seed}{Seed for reproducibility of posterior prediction.}

\item{...}{Parameters passed to \code{\link[projpred:project]{projpred::project()}}.}
}
\value{
List containing:
\describe{
\item{model}{Projected model}
\item{ppred}{Posterior predictions from the model}
\item{probs}{Expected posterior probabilities for data used to fit the model}
}
}
\description{
Given a vsel variable selection object and size of the projected model,
projects posterior of the reference model onto the selected submodel
parameter space.
}
