% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_varsel.R
\name{perform_varsel}
\alias{perform_varsel}
\title{Perform variable selection}
\usage{
perform_varsel(refm_fit, ...)
}
\arguments{
\item{refm_fit}{Reference model, e.g., fitted via \code{fit_reference}.}

\item{...}{Parameters passed to \code{\link[projpred:cv_varsel]{projpred::cv_varsel()}}.}
}
\value{
An object of class vsel as returned by \code{\link[projpred:cv_varsel]{projpred::cv_varsel()}}.
}
\description{
A wrapper for \code{\link[projpred:cv_varsel]{projpred::cv_varsel()}}. Using a provided reference model,
runs a projective prediction feature selection via cross-validation.
}
\seealso{
\itemize{
\item \code{\link[=fit_reference]{fit_reference()}}
\item \code{\link[projpred:cv_varsel]{projpred::cv_varsel()}}
}
}
