% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_rates.r
\docType{data}
\name{rates}
\alias{rates}
\title{Algorithm Specifications and PDD Rates}
\format{
A tibble with 68 rows and 11 columns:
\describe{
\item{type}{Character; name of the algorithm.}
\item{perc}{Numeric; raw PDD rate.}
\item{Global}{Character; operationalisation of global cognitive deficit.}
\item{Executive}{Character; operationalisation of executive deficit.}
\item{Construction}{Character; operationalisation of visuoconstruction deficit.}
\item{Memory}{Character; operationalisation of memory deficit.}
\item{Language}{Character; operationalisation of language deficit.}
\item{IADL}{Character; operationalisation of IADL deficit.}
\item{N}{Integer; number of observations for a given algorithm.}
\item{Rate}{Character; number and percentage of patients with probable PDD.}
}
}
\source{
Generated by the targets pipeline in \verb{_targets.R}.
}
\usage{
rates
}
\description{
A comprehensive table containing all algorithms evaluated in the article. It
containing algorithm names as well as their constituting components including
operationalisations of IADL deficit, global cognitive deficit and cognitive
impairment according to monitored domains.
}
\seealso{
\link{concords} for data containing pairwise concordance statistics.
}
\keyword{datasets}
