% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_pdd_on_demographics.R
\name{regress_pdd_on_demographics}
\alias{regress_pdd_on_demographics}
\title{Regress PDD via Logistic Regression}
\usage{
regress_pdd_on_demographics(d0, d1, covs = NULL, inter = TRUE)
}
\arguments{
\item{d0}{A tibble of demographic data generated by \code{prepare_data}.}

\item{d1}{A tibble of PDD diagnoses (typically the \code{PDD} element from
the list returned by \code{diagnose_pdd_sample}).}

\item{covs}{Either a character vector containing covariates of interest (in
which case the resulting formula will be \code{PDD ~ age * sex * (covariate1 + covariate2 + ...)},
or \code{NULL} (default case whereby the resulting formula will be \code{PDD ~ age * sex}).}

\item{inter}{A logical indicating whether covariates ought to be added to the
together with their interactions with predictors of interest (default and
recommended, \code{TRUE}) or as additive terms only (\code{FALSE}).}
}
\value{
A list with the following components:
\describe{
\item{\code{fits}}{A list of fitted logistic regression models for each PDD
criterion.}
\item{\code{values}}{A tibble of model coefficients and statistics.}
\item{\code{plots}}{A named list of ggplot2 objects:
\describe{
\item{\code{data}}{Plots showing raw proportions of PDD diagnoses by
age and sex.}
\item{\code{parameters}}{Plots summarising effect sizes (odds ratios)
and associated p-values.}
}}
}
}
\description{
For each algorithm for probable Parkinson’s disease dementia (PDD), this
function fits a logistic regression model predicting the diagnosis from age,
sex, and their interaction. It takes as input demographic data and a tibble
of PDD diagnoses (typically generated via \code{diagnose_pdd_sample}).
}
\examples{
\dontrun{
p <- data_paths("data-raw")
data0 <- prepare_data(p)
data1 <- diagnose_pdd_sample(data0)
results <- regress_pdd_on_demographics(data0, data1$PDD)
}

}
\seealso{
\itemize{
\item \code{\link[=prepare_data]{prepare_data()}} prepares \code{d0}.
\item \code{\link[=diagnose_pdd_sample]{diagnose_pdd_sample()}} prepares \code{d1}.
\item \code{\link[stats:glm]{stats::glm()}} is used to fit regressions.
}
}
