% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_roc.R
\name{run_roc}
\alias{run_roc}
\title{Run ROC analysis}
\usage{
run_roc(y_obs, preds, prevs = seq(0.1, 0.5, 0.1))
}
\arguments{
\item{y_obs}{Observed values}

\item{preds}{Predictions. Does not need to be on the same scale as observed
values. Can be single test scores or expected probabilities from a logistic
regression or any other model.}

\item{prevs}{Prevalences.}
}
\value{
List with two components:
\describe{
\item{ROC}{List of class "roc", including AUC plot.}
\item{thresholds}{Tibble with the best thresholds according to assumed PDD
prevalences with additional columns for specificity, sensitivity, and
accuracy in current sample.}
}
}
\description{
Given observed values and predictions, runs ROC analysis.
}
