% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scoring_rule_pipeline.R
\name{run_scoring_rule_pipeline}
\alias{run_scoring_rule_pipeline}
\title{Run pipeline deriving scoring rule for PDD}
\usage{
run_scoring_rule_pipeline(model_grid, ..., nms = NULL)
}
\arguments{
\item{model_grid}{A single row of \code{grid_models}.}

\item{...}{Other parameters going either to \code{run_roc} if \code{prevs} or
to \code{extract_coefficients} if \code{stat}.}

\item{nms}{Optional coefficient names, including intercepts. Double check
they are in correct order!}
}
\value{
description
}
\description{
Using different assumed PDD prevalences and different models, computes a
set of ROC analyses and derives decision rules for the best threshold
identified by each.
}
