% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_criteria.R
\name{specify_criteria}
\alias{specify_criteria}
\title{Specify criteria.}
\usage{
specify_criteria()
}
\value{
A tibble with columns indicating
type of criterion as well as single indexes
and thresholds. To be used within a wrapper
function \code{diagnose_pdd_sample}.
}
\description{
This is purely a helper function that
specifies a tibble with specifications
of single algorithms for PDD diagnosis.
}
