% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract_concordance_matrixes.R
\name{subtract_concordance_matrixes}
\alias{subtract_concordance_matrixes}
\title{Compare concordance matrixes}
\usage{
subtract_concordance_matrixes(conc0, conc1, label0 = "0", label1 = "1")
}
\arguments{
\item{conc0}{The subtrahend concordance statistics}

\item{conc1}{The minuend concordance statistics}

\item{label0}{How should the subtrahend be called?}

\item{label1}{How should the minuend be called?}
}
\value{
A list with two components:
\describe{
\item{\code{differences}}{Tibble containing rowwise differences}.
\item{\code{plots}}{A list of ggplot2-based visualisations of differences}
}
}
\description{
Given two sets of \code{describe_concordance} results, computes difference
matrixes for Cohen's Kappa, Accuracy, Sensitivity, and Specificity.
}
