% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_kappa.r
\name{summarise_kappa}
\alias{summarise_kappa}
\title{Summarise Kappa Coefficients}
\usage{
summarise_kappa(algorithms, concordance)
}
\arguments{
\item{algorithms}{A list containing algorithm mapping generated by
\code{list_algorithms}.}

\item{concordance}{A tibble containing concordance indexes for pairs
of algorithms generated by \code{decribe_concordance}. The table
is also available as data \link{concords}.}
}
\value{
A list with the following components:
\describe{
\item{\code{faq_tot}}{A numeric vector containing all Cohen's kappas
of algorithm pairs with IADL deficit defined as FAQ total > 7.}
\item{\code{faq_9}}{A numeric vector containing all Cohen's kappas
of algorithm pairs with IADL deficit defined as FAQ item 9 > 1.}
\item{\code{cross}}{A numeric vector containing all Cohen's kappas
of algorithm pairs with different IADL deficit definitions.}
\item{\code{iadl}}{A numeric vector containing all Cohen's kappas
of algorithm pairs which differ only in IADL deficit definition.}
\item{\code{sums}}{A named character vector containing mean and
standard deviation of each vector of Kappas contained in this list.}
}
}
\description{
Using a list of PDD algorithms stratified by IADL deficit operationalisation
and a table of pairwise concordance statistics, this function extracts the
raw Cohen’s kappa estimates and computes their summary statistics (mean and
standard deviation) across groups of algorithm pairs defined by the IADL
operationalisation scheme.
}
\examples{
\dontrun{
p <- data_paths("data-raw")
data <- prepare_data(p)
pdd  <- diagnose_pdd_sample(data)
vars <- here::here("data-raw", "VariablesOfInterest.csv")

data(concords)
rates <- summarise_rates(pdd, vars)
algos <- list_algorithms(rates$table)

# Two equally valid ways to run the function:
kappa1 <- summarise_kappa(algos, rates$table)
kappa2 <- summarise_kappa(algos, concords)
}

}
\seealso{
\itemize{
\item \code{\link[=list_algorithms]{list_algorithms()}} prepares \code{algorithms}.
\item \code{\link[=describe_concordance]{describe_concordance()}} prepares \code{concordance}.
\item \link{concords} containg data used as \code{concordance}.
}
}
