% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_prevalence.R
\name{summarise_prevalence}
\alias{summarise_prevalence}
\title{Orders and prints PDD operationalisations based
on their estimated prevalence.}
\usage{
summarise_prevalence(d0, vars, descending = T)
}
\arguments{
\item{d0}{A list with PDD data generated via
\code{diagnose_pdd_sample}.}

\item{vars}{A data.frame/tibble/matrix with variable
names in the first column, variable labels in the
second column, type of variable (continuous, binary or
nominal) in the third column, and optionally a fourth column
denoting group and a fifth column that maps each
label to its description in the table's note.
Alternatively, a path to a csv delimited by semi-colon
containing such a table.}

\item{descending}{A logical indicating whether
operationalisations with the highest estimated
PDD prevalence should be on the top (by default
descending = T).}
}
\value{
A list with a tibble containing
raw summaries, a APA-style gt table and
a ggplot summarising prevalence estimates.
}
\description{
This function computes summaries of the
prevalence estimates stratified by PDD
operationalisation.
}
\examples{
\dontrun{
p    <- data_paths('data-raw')
data <- prepare_data(p)
pdd  <- diagnose_pdd_sample(data)
vars <- here::here('data-raw','VariablesOfInterest.csv')

prevtab0 <- table_prevalence(pdd, vars)
prevtab1 <- table_prevalence(pdd, vars, T) # prevtab0 & prevtab1 are identical
prevtab2 <- table_prevalence(pdd, vars, F)
}
}
