% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_algorithms.R
\name{table_algorithms}
\alias{table_algorithms}
\title{Generate a Summary Table of Algorithms}
\usage{
table_algorithms(vars)
}
\arguments{
\item{vars}{A data.frame, tibble, or matrix with  in the following order:
\enumerate{
\item variable names,
\item variable labels
\item type of variable (continuous, binary, or nominal)
\item optional, group,
\item optional, mapping each label to its description in the table’s note.
}

Alternatively, a path to a CSV file (semicolon-delimited) containing such
a table.}
}
\value{
A \code{gt} table object containing the summary of criteria used
in the study.
}
\description{
This function creates a pre-specified summary table of the probable PDD
algorithms compared in the study. It is intended as a shortcut to insert
a clean table into a Quarto document without cluttering the main text.
The table content is hard-coded and can only be changed by modifying the
function code.
}
\seealso{
\code{\link[=gt_apa_table]{gt_apa_table()}} is used to format the table.
}
