% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_levelII_approximations.R
\name{table_levelII_approximations}
\alias{table_levelII_approximations}
\title{Summary Table of Predictors}
\usage{
table_levelII_approximations(values, struct, crit)
}
\arguments{
\item{values}{A tibble of concordance statistics generated by
\code{describe_concordance}. Can also be provided by the data \link{concords}
propagated with the package.}

\item{struct}{A list generated by \code{list_algorithms} with two elements:
\describe{
\item{\verb{[[1]]}}{Character vector of Level I algorithm names derived using
FAQ total > 7 as the IADL criterion.}
\item{\verb{[[2]]}}{Character vector of Level I algorithm names derived using
FAQ item 9 > 1 as the IADL criterion. Note: order matters.}
}}

\item{crit}{A character string indicating which concordance statistics is to
be used for comparisons. Can be one of \code{"Accuracy_raw"}, \code{"Balanced Accuracy"},
\code{"Kappa_raw"}, \code{"Sensitivity"} or \code{"Specificity"}.}
}
\value{
A list containing:
\describe{
\item{\code{table}}{A tibble with the five best and five worst Level I
algorithms in predicting the specified Level II-based diagnosis of
probable PDD.}
\item{\code{gtable}}{A \code{gt} table object summarising the concordance
statistics (e.g., accuracy, kappa) of the five best and five worst Level I
algorithms in predicting the specified Level II-based diagnosis of probable
PDD.}
}
}
\description{
This helper function prepares a table showing the five best and five
worst-performing Level I algorithms in predicting a selected Level II-based
diagnosis of probable PDD. It is primarily intended for use in the manuscript
to keep Quarto document uncluttered.
}
\examples{
\dontrun{
p <- data_paths("data-raw")
data <- prepare_data(p)
pdd  <- diagnose_pdd_sample(data)
vars <- here::here("data-raw", "VariablesOfInterest.csv")
rates <- summarise_rates(pdd, vars)
algos <- list_algorithms(rates$table)
data(concords)

# Tables ordered by different statistics:
table_levelII_approximations(concords, algos, "Accuracy_raw") # Table 3
table_levelII_approximations(concords, algos, "Balanced Accuracy") # Table 7
table_levelII_approximations(concords, algos, "Kappa_raw")
table_levelII_approximations(concords, algos, "Sensitivity")
table_levelII_approximations(concords, algos, "Specificity")
}

}
\seealso{
\itemize{
\item \link{concords} contains data for \code{values}.
\item \code{\link[=describe_concordance]{describe_concordance()}} prepares \code{values}.
\item \code{\link[=list_algorithms]{list_algorithms()}} prepares \code{struct}.
\item \code{\link[=gt_apa_table]{gt_apa_table()}} formats the table.
\item \code{\link[psych:kappa]{psych::cohen.kappa()}} computes and documents Cohen’s kappa and related metrics.
\item \code{\link[caret:confusionMatrix]{caret::confusionMatrix()}} computes and documents the remaining metrics for \code{crit}.
}
}
