% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_data.R
\name{trim_data}
\alias{trim_data}
\title{Prepare complete data}
\usage{
trim_data(d0, inc)
}
\arguments{
\item{d0}{Data to be trimmed}

\item{inc}{Incomplete cases, e.g., generated by \code{find_incomplete}}
}
\value{
The same data as input but with complete cases (regarding /PDD
data) only.
}
\description{
Given a set o patients with incomplete PDD data, extracts data set containing
only complete cases. Can be applied to raw (generated by \code{prepare_data})
or pre-processed (generated by \code{diagnose_pdd_sample}) PDD data.
}
