% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_compatibility.R
\name{check_compatibility}
\alias{check_compatibility}
\title{Check Data Compatibility}
\usage{
check_compatibility(d1, d2)
}
\arguments{
\item{d1}{A data frame generated by \code{import_item_data}, containing item-level data.}

\item{d2}{A data frame generated by \code{import_redcap_data}, containing REDCap data.}
}
\value{
If discrepancies are found, an informative message is printed and the function
stops execution (e.g., via \code{stop()}). This function is intended to be used
inside \code{prepare_data()} as a consistency check.
}
\description{
Compares patient names between the hand-entered item-level data and the REDCap export
to identify any mismatches or inconsistencies.
}
\examples{
\dontrun{
p <- data_paths("data-raw")
data <- prepare_data(p)  # Internally calls the compatibility check
}
}
\seealso{
\itemize{
\item \code{\link[=prepare_data]{prepare_data()}} is a wrapper of this function.
\item \code{\link[=import_item_data]{import_item_data()}} prepares \code{d1}.
\item \code{\link[=import_redcap_data]{import_redcap_data()}} prepares \code{d2}.
}
}
