% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_compatibility.R
\name{check_compatibility}
\alias{check_compatibility}
\title{Check compatibility between item and REDCap data.}
\usage{
check_compatibility(d1, d2)
}
\arguments{
\item{d1}{Item-wise data set generated by \code{import_item_data}.}

\item{d2}{REDCap data set generated by \code{import_redcap_data}.}
}
\value{
Nothing, only prints a message with data
discrepancies listed and stops the analysis if there
are any. Ought to be run within the \code{prepare_data}.
}
\description{
This function compares patients' names
in the hand-written data and REDCap.
}
\examples{
\dontrun{
p    <- data_paths('data-raw')
data <- prepare_data(p)
}
}
