% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_names.R
\name{check_names}
\alias{check_names}
\title{Check names of the hand-written data.}
\usage{
check_names(d, nms)
}
\arguments{
\item{d}{Data set with names to be checked.}

\item{nms}{Data set with proper names according
to REDCap.}
}
\value{
Nothing, only prints a message with names
discrepancies. Ought to be run within the \code{prepare_data}
function where it could be made silent.
}
\description{
This function compares patients' names
in the hand-written data and REDCap.
}
\examples{
\dontrun{
p    <- data_paths('data-raw')

# Checks names:
data <- prepare_data(p)
data <- prepare_data(p, check.names = T)

# Does not check names:
data <- prepare_data(p, check.names = F)
}
}
