% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ranges.R
\name{check_ranges}
\alias{check_ranges}
\title{Check ranges of included data.}
\usage{
check_ranges(d)
}
\arguments{
\item{d}{A data set prepared within the
\code{prepare_data} function.}
}
\value{
Optionally a message with data that
are out of range and indicator to stop the
analysis.
}
\description{
This checks data ranges of data for the
analysis.
}
\examples{
\dontrun{
p    <- data_paths('data-raw')
data <- prepare_data(p)
}
}
