% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_descriptives.R
\name{compute_descriptives}
\alias{compute_descriptives}
\title{Describe the Sample}
\usage{
compute_descriptives(d0, vois)
}
\arguments{
\item{d0}{A tibble containing the data, as generated by \code{prepare_data}.}

\item{vois}{A data frame, tibble, or matrix with the following columns (in this order):
\enumerate{
\item variable name,
\item variable label,
\item variable type (one of "continuous", "binary", "nominal"),
\item optional: variable group,
\item optional: label-to-description mapping for table notes,
\item optional: score type description to append in parentheses.
}

Alternatively, a path to a CSV file (semicolon-delimited) containing this information.}
}
\value{
A list containing:
\describe{
\item{\code{table}}{A tibble with raw descriptive statistics.}
\item{\code{gtable}}{An APA-style formatted {gt} table summarizing selected variables.}
}
}
\description{
Computes descriptive summaries for a specified set of variables in a dataset.
The output includes a raw summary tibble and an APA-style formatted gt table.
}
\examples{
\dontrun{
p <- data_paths("data-raw")
data <- prepare_data(p)
demo <- compute_descriptives(data, here::here("data-raw", "VariablesOfInterest.csv"))
}

}
\seealso{
\code{\link[=prepare_data]{prepare_data()}} prepares \code{d0}.
}
