% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_descriptives.R
\name{compute_descriptives}
\alias{compute_descriptives}
\title{Compute table with description of sample variables.}
\usage{
compute_descriptives(d0, vois)
}
\arguments{
\item{d0}{Data (tibble) as generated by
\code{prepare_data}.}

\item{vois}{A data.frame/tibble/matrix with variable
names in the first column, variable labels in the
second column, type of variable (continuous, binary or
nominal) in the third column, and optionally a fourth column
denoting group and a fifth column that maps each
label to its description in the table's note.
Alternatively, a path to a csv delimited by semi-colon
containing such a table.}
}
\value{
A list with a tibble containing
raw summaries, APA-style gt table and a
(maybe in a future) a set of ggplots.
}
\description{
This function computes summaries of the
sample (tables, in the future maybe including
their graphical representation).
}
\examples{
\dontrun{
p    <- data_paths('data-raw')
data <- prepare_data(p)
demo <- compute_descriptives(data, here::here('data-raw','VariablesOfInterest.csv'))
}
}
