% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_paths.R
\name{data_paths}
\alias{data_paths}
\title{Define paths to data.}
\usage{
data_paths(dir = "data-raw")
}
\arguments{
\item{dir}{Input directory where data reside,
defaults to dir = 'data-raw'.}
}
\value{
A character vector with paths towards
all raw data used in this study.
}
\description{
This function prepares a set of paths to data
that will be used in the analysis. The naming
conventions are set and the function will inform
the user if the data in the folder of interest ain't
compatible with the script.
To be then pushed to \code{prepare_data} function
for pre-processing.
}
\examples{
\dontrun{
p    <- data_paths('data-raw')
data <- prepare_data(p)
}
}
