% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_concordance.R
\name{describe_concordance}
\alias{describe_concordance}
\title{Compute Concordance Measures}
\usage{
describe_concordance(d0)
}
\arguments{
\item{d0}{A list containing PDD diagnostic results, as produced by
\code{diagnose_pdd_sample}.}
}
\value{
A list with two components:
\describe{
\item{\code{table}}{A data frame summarising pairwise concordance statistics
(e.g., Kappa, Accuracy, Sensitivity, Specificity, PPV, NPV). The table is propagated
with the package as data \code{concords}}.
\item{\code{plots}}{A list of ggplot2-based visualisations, including matrices
of Cohen's Kappa, Accuracy (with significance tests against the No Information Rate),
Sensitivity, and Specificity.}
}
}
\description{
Computes pairwise concordance statistics between different algorithms for
probable Parkinson’s disease dementia (PDD). Metrics include Cohen's Kappa,
Sensitivity, Positive Predictive Value (PPV), Specificity, Negative Predictive
Value (NPV), and Accuracy. Visual summaries are also generated via {ggplot2}.
}
\examples{
\dontrun{
p <- data_paths("data-raw")
data <- prepare_data(p)
pdd <- diagnose_pdd_sample(data)
concmats <- describe_concordance(pdd)
}

}
\seealso{
\itemize{
\item \code{\link[=diagnose_pdd_sample]{diagnose_pdd_sample()}} prepares \code{d0}.
\item \link{concords} contains the table of results generated by this function.
\item \code{\link[psych:kappa]{psych::cohen.kappa()}} computes and documents Cohen’s kappa and related
metrics.
\item \code{\link[caret:confusionMatrix]{caret::confusionMatrix()}} computes and documents the remaining metrics.
}
}
