% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_concordance.R
\name{describe_concordance}
\alias{describe_concordance}
\title{Compute and visualise measures of concordance between diagnostic criteria}
\usage{
describe_concordance(d0, vars)
}
\arguments{
\item{d0}{A list containing PDD diagnostic results, as produced by
\code{diagnose_pdd_sample}.}
}
\value{
A list with two components:
\describe{
\item{\code{stats}}{A data frame summarising pairwise concordance statistics
(e.g., Kappa, Accuracy, Sensitivity, Specificity, PPV, NPV).}
\item{\code{plots}}{A list of ggplot2-based visualisations, including matrices
of Cohen's Kappa, Accuracy (with significance tests against the No Information Rate),
Sensitivity, and Specificity.}
}
}
\description{
Computes pairwise concordance statistics between different operationalisations
of probable Parkinson’s disease dementia (PDD). Metrics include Cohen's Kappa,
Sensitivity, Positive Predictive Value (PPV), Specificity, Negative Predictive
Value (NPV), and Accuracy. Visual summaries are also generated.
}
\examples{
\dontrun{
p <- data_paths("data-raw")
data <- prepare_data(p)
pdd <- diagnose_pdd_sample(data)
concmats <- describe_concordance(pdd)
}
}
