% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_pdd_case.R
\name{diagnose_pdd_case}
\alias{diagnose_pdd_case}
\title{Diagnose PDD of a sample of patients using
one set of criteria.}
\usage{
diagnose_pdd_case(x, c)
}
\arguments{
\item{x}{A patients data matrix.}

\item{c}{A vector with PDD
criteria.}
}
\description{
This function takes in a data matrix of
patients' data and a vector of criteria to
use and diagnoses each patient with PDD
(or not) according to the criteria. Ought
to be used within the wrapper function
\code{diagnose_pdd_sample}.
}
\examples{
\dontrun{
p    <- data_paths('data-raw')
data <- prepare_data(p)
crit <- specify_criteria()
pdd  <- diagnose_pdd_case(data, crit[1, ])
}.

}
