% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_pdd_case.R
\name{diagnose_pdd_case}
\alias{diagnose_pdd_case}
\title{Diagnose via a Single Algorithm}
\usage{
diagnose_pdd_case(x, c)
}
\arguments{
\item{x}{A data frame or matrix containing patient-level variables used for diagnosis.}

\item{c}{A named vector (or single-row data frame) specifying the criteria
to be applied for PDD diagnosis.}
}
\value{
A logical vector indicating probable PDD diagnosis for each
patient (\code{TRUE} = diagnosed, \code{FALSE} = not diagnosed).
}
\description{
Applies a specified set of diagnostic criteria to a data matrix representing
patient-level data to determine whether each patient meets the criteria for
probable Parkinson’s disease dementia (PDD).
This function is intended to be used internally by the wrapper function
\code{diagnose_pdd_sample}.
}
\examples{
\dontrun{
p <- data_paths("data-raw")
data <- prepare_data(p)
crit <- specify_criteria()
pdd  <- diagnose_pdd_case(data, crit[1, ])
}

}
\seealso{
\code{\link[=diagnose_pdd_sample]{diagnose_pdd_sample()}} wraps the function to diagnose all patients.
}
