% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_pdd_sample.R
\name{diagnose_pdd_sample}
\alias{diagnose_pdd_sample}
\title{Diagnose PDD using all available criteria
in all patients.}
\usage{
diagnose_pdd_sample(d0)
}
\arguments{
\item{d0}{A tibble with the data generated
by \code{prepare_data}}
}
\value{
A list containing two tibbles with
(i) criteria specification and (ii) patient id,
criterion type (to be mapped to criteria specification),
PDD diagnosis, single criteria akin to Table 2 of
Dubois et al. (2007) and impairment statements for
each cognitive domain according to Level I.
}
\description{
This function takes in study data and loops
through all criteria specifications (generated
internally via the \code{specify_criteria}
function) and patients to arrive at PDD
diagnosis.
}
\examples{
\dontrun{
p    <- data_paths('data-raw')
data <- prepare_data(p)
pdd  <- diagnose_pdd_sample(data)
}

}
