% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_summary.R
\name{do_summary}
\alias{do_summary}
\title{Summarises data in one of a set of pre-specified ways.}
\usage{
do_summary(x, dec, sum = NULL)
}
\arguments{
\item{x}{A vector to be summarised.}

\item{dec}{An integer denoting decimals.}

\item{sum}{A type of summary. Takes in NULL (default)
for simple printing, 'N' for number of observations,
'msd' for M ± SD, 'M' for mean, SD' for standard deviation,
'Md' for median, 'IQR', for interquartile range,
'minmax' for minimum-maximum, Min' for minimum, 'Max'
for maximum, p' for a p-value (for tables), 'ptext'
for a p-value (for in-text), 'Nperc' for a N (\%)
representation of a binary variable, 'Nslash' for a
N(group1)/N(group2)/... representation of nominal
variable with unlimited number of categories.}
}
\value{
A summary of choice.
}
\description{
This function computes demographic summaries
of the sample including their graphical
representation.
}
\examples{
\dontrun{
p    <- data_paths('data-raw')
data <- prepare_data(p)

M  <- do_summary(data[ , 25], 2, 'M')
SD <- do_summary(data[ , 25], 2, 'SD')
Md <- do_summary(data[ , 25], 2, 'Md')
mm <- do_summary(data[ , 25], 0, 'minmax')
}
}
