% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_models.R
\name{grid_models}
\alias{grid_models}
\title{Fit a grid of reference models}
\usage{
grid_models(d, gss = c("Lvl.II (1)", "Lvl.II (2)"), N = 5, ...)
}
\arguments{
\item{d}{Input data, e.g., generated by \code{diagnose_pdd_sample}.}

\item{gss}{Gold standard variables.}

\item{N}{Vector of expected non-zero coefficient for a regularised horseshoe
prior.}

\item{...}{Other parameters for \code{\link[=fit_reference]{fit_reference()}}.}
}
\value{
A grid with columns:
\describe{
\item{gs}{Gold standard used.}
\item{esig}{Number of expected non-zero parameters.}
\item{fit}{Model data as generated by \code{fit_reference}}
}
}
\description{
Using different values of "gold standard" measures of the outcome and number
of expected non-zero parameters, fits a set of reference models via
\code{fit_reference}.
}
\seealso{
\code{\link[=fit_reference]{fit_reference()}}
}
