% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_postpred.R
\name{grid_postpred}
\alias{grid_postpred}
\title{Grid posterior predictions of a model}
\usage{
grid_postpred(
  mod,
  pgrid = NULL,
  scls = NULL,
  perc = 0.95,
  lt = c(-1.29335, -2.649892),
  cols = c("grey80", "black", "red4"),
  pal = "viridis",
  dir = 1
)
}
\arguments{
\item{mod}{brms or projpred model}

\item{pgrid}{A grid of predictor values.}

\item{scls}{A scaling table with columns "x" for predictor name, "M" for mean
and "SD" for standard deviation.}

\item{perc}{Percentage value for computation of the equal-tailed posterior
interval \(ETI\).}

\item{lt}{Logit threshold(s) for declaring PDD.}

\item{cols}{Text colours for separating PDD classification according to
thresholds in \code{lt}. Must be of length \code{length(lt) + 1}}

\item{pal}{Colour pallete pushed to \code{\link[ggplot2:scale_viridis]{ggplot2::scale_fill_viridis_c()}}.}

\item{dir}{Which way should be \code{pal} scaled? Pushed to
\code{\link[ggplot2:scale_viridis]{ggplot2::scale_fill_viridis_c()}} as \code{direction}.}
}
\value{
\{ggplot2\} heatmap with estimates and \code{perc}\% ETIs.
}
\description{
Given a grid of predictor values, their scaling, and a Bayesian logistic
regression model, prints a prediction grid heatmap.
}
