% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_apa_table.R
\name{gt_apa_table}
\alias{gt_apa_table}
\title{Get APA-style gt table.}
\usage{
gt_apa_table(x, grp = NULL, nms = NULL, tit = "")
}
\arguments{
\item{x}{A tibble/data.frame to be APAed.}

\item{grp}{A column name to be applied to
the gt::gt() groupname_col parameter (NULL
by default).}

\item{nms}{A column name to be applied to
the gt::gt() rowname_col parameter (NULL
by default).}

\item{tit}{A character including table name.
Will be forced to become html (No Name by default).}
}
\value{
A gt object of a table in APA format.
}
\description{
This function takes in a tibble or a data.frame
and returns APA-style table created via the gt
package. All gt quirks and perks apply to the
resulting object.
}
