% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_item_data.R
\name{import_item_data}
\alias{import_item_data}
\title{Import item-wise data.}
\usage{
import_item_data(path)
}
\arguments{
\item{paths}{Path to data. Should be generated
by \code{data_paths}.}
}
\value{
A tibbles with item level data.
}
\description{
Using data path, this function imports the
item-level data in a suitable format.
It also checks for unrealistic values in MMSE.
The function and is supposed to be run
within the larger \code{prepare_data}.
}
\examples{
\dontrun{
p    <- data_paths('data-raw')
data <- import_item_data(p[1])
}
}
