% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_redcap_data.R
\name{import_redcap_data}
\alias{import_redcap_data}
\title{Import REDCap data.}
\usage{
import_redcap_data(path, scoring)
}
\arguments{
\item{paths}{Path to data. Should be generated
by \code{data_paths}.}
}
\value{
A tibbles with appropriate REDCap data.
}
\description{
Using data path and scoring information,
this function imports the REDCap data in
a suitable format. The function and is
supposed to be run within the larger
\code{prepare_data}.
}
\examples{
\dontrun{
p    <- data_paths('data-raw')
scor <- readr::read_delim(p[4], delim = ";")
data <- import_redcap_data(p[2])
}

}
