% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_redcap_data.R
\name{import_redcap_data}
\alias{import_redcap_data}
\title{Import REDCap data}
\usage{
import_redcap_data(path, scoring)
}
\arguments{
\item{paths}{A character string specifying the path to the REDCap data file.
Typically generated by the \code{data_paths} function.}
}
\value{
A tibble containing the imported and processed REDCap data.
}
\description{
Imports REDCap dataset from a specified file path using scoring information.
This function prepares the data in a suitable format and is intended to be used
within the broader \code{prepare_data} workflow.
}
\examples{
\dontrun{
p <- data_paths("data-raw")
scor <- readr::read_delim(p[4], delim = ";")
data <- import_redcap_data(p[2])
}

}
