% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_redcap_data.R
\name{import_redcap_data}
\alias{import_redcap_data}
\title{Import REDCap Data}
\usage{
import_redcap_data(path, scoring)
}
\arguments{
\item{path}{A character string specifying the path to the REDCap data file.
Typically generated by the \code{data_paths} function.}

\item{scoring}{A tibble containing scoring conventions for variables of
interest.}
}
\value{
A tibble containing the imported and processed REDCap data.
}
\description{
Imports REDCap dataset from a specified file path using scoring information.
This function prepares the data in a suitable format and is intended to be used
within the broader \code{prepare_data} workflow.
}
\examples{
\dontrun{
p <- data_paths("data-raw")
scor <- readr::read_delim(p[4], delim = ";", col_types = readr::cols(rev = "c"))
data <- import_redcap_data(p[2])
}

}
\seealso{
\itemize{
\item \code{\link[=data_paths]{data_paths()}} prepare \code{path}.
\item \code{\link[=prepare_data]{prepare_data()}} is a wrapper of this function.
}
}
