% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dendrogram.r
\name{make_dendrogram}
\alias{make_dendrogram}
\title{Make a Dendrogram of PDD Algorithms}
\usage{
make_dendrogram(tab)
}
\arguments{
\item{tab}{A table containing all criteria used as generated by
\code{summarise_rates}.}
}
\value{
A ggplot object with a dendrogram created
by {ggraph}.
}
\description{
The functions prepares a dendrogram (tree plot) that summarises PDD algorithms
construction starting from IADL operationalisation at the top followed successively
by operationalisation of Global cognitive performance, Attention, Executive
Function, Construction, Memory and Language. The dendrogram visualises PDD rate
estimated by each algorithm via point size.
}
\seealso{
\code{\link[=summarise_rates]{summarise_rates()}} which is fed into this function.
}
