% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlations.R
\name{plot_correlation}
\alias{plot_correlation}
\title{Plot correlation matrixes}
\usage{
plot_correlation(corrs)
}
\arguments{
\item{corrs}{List with correlation matrixes.}
}
\value{
Plots the correlation matrixes from \code{corrs} side by side;
with left (A) depicting observed correlations, right (B) depicting the
"ideal" correlation matrix.
}
\description{
Generates two correlation matrixes via \code{\link[corrplot:corrplot]{corrplot::corrplot()}} and plots
them side to side. Intended to be used for visual comparison of observed
and "ideal" correlations as generated by \code{correlate_indexes}. Format
is fixed for purposes of \{demcrit\}.
}
