% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Pre-process data.}
\usage{
prepare_data(p, check.names = T)
}
\arguments{
\item{p}{Paths to data. Should be generated
by \code{data_paths}.}
}
\value{
A tibble including relevant variables.
It also prints information regarding potential
discrepancies in data and stops the analysis
in some cases.
}
\description{
Using data paths, this functions check their
format and compatibility and if compatible,
the function proceeds to pre-process data.
}
\examples{
\dontrun{
p    <- data_paths('data-raw')
data <- prepare_data(p)
}
}
