% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_defaults.r
\name{prepare_defaults}
\alias{prepare_defaults}
\title{Prepare Defaults for Data Simulation}
\usage{
prepare_defaults()
}
\value{
A list containing:
\describe{
\item{\code{Mu}}{Named numeric vector containing means for FAQ, MMSE, MoCA,
and sMoCA.}
\item{\code{Sigma}}{Named matrix containing variance-covariance of the variables
(FAQ, MMSE, MoCA, and sMoCA).}
\item{\code{cens}}{A matrix with named rows denoting scales (FAQ, MMSE, MoCA,
and sMoCA) and two columns - the first containing lower bound, the second
containing upper bound of each scale.}
\item{\code{crits}}{A data.frame with four columns denoting IADL measure to use
(\code{IADL}), threshold for the IADL measure (\code{IADL_thres}), cognitive measure to
use (\code{cognition}), and threshold for the cognitive measure (\code{cognition_thres}).
Score above \code{IADL_thres} and below \code{cognition_thres} are regarder to indicate
PDD symptoms. Rownames contain algorithm names.}
}
}
\description{
Prepares a list with default values for data simulation. These defaults are
derived from the data used as a basis of the demcrit article. This function
is a helper with no inputs and is typically called within \code{simulate_PDD_data}.
}
\seealso{
\code{\link[=simulate_pdd_data]{simulate_pdd_data()}} uses outputs of this function unless specified
by the user.
}
