% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_pdd_on_demographics.R
\name{regress_pdd_on_demographics}
\alias{regress_pdd_on_demographics}
\title{Regress probable PDD on age and sex across diagnostic criteria}
\usage{
regress_pdd_on_demographics(d0, d1)
}
\arguments{
\item{d0}{A tibble of demographic data generated by \code{prepare_data}.}

\item{d1}{A tibble of PDD diagnoses (typically the \code{PDD} element from
the list returned by \code{diagnose_pdd_sample}).}
}
\value{
A list with the following components:
\describe{
\item{\code{fits}}{A list of fitted logistic regression models for each PDD criterion.}
\item{\code{values}}{A tibble of model coefficients and statistics.}
\item{\code{plots}}{A named list of ggplot2 objects:
\describe{
\item{\code{data}}{Plots showing raw proportions of PDD diagnoses by age and sex.}
\item{\code{parameters}}{Plots summarising effect sizes (odds ratios) and associated p-values.}
}}
}
}
\description{
For each operationalisation of probable Parkinson’s disease dementia (PDD),
this function fits a logistic regression model predicting diagnosis from age,
sex, and their interaction. It takes as input demographic data and a tibble
of PDD diagnoses (typically generated via \code{diagnose_pdd_sample}).
}
\examples{
\dontrun{
p <- data_paths("data-raw")
data0 <- prepare_data(p)
data1 <- diagnose_pdd_sample(data0)
results <- regress_pdd_on_demographics(data0, data1$PDD)
}

}
