% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_algorithms.R
\name{specify_algorithms}
\alias{specify_algorithms}
\title{Specify Algorithms for Probable PDD}
\usage{
specify_algorithms()
}
\value{
A tibble where each row represents one unique diagnostic algorithm
for probable PDD. Columns indicate the type of algorithms, individual test
indexes, thresholds, and other relevant parameters.
}
\description{
Computes a Cartesian product of possible diagnostic components to define
all combinations of algorithms for probable Parkinson’s disease dementia (PDD).
This function is a helper with no inputs and is typically called within
\code{diagnose_pdd_sample}.
}
\seealso{
\code{\link[=diagnose_pdd_sample]{diagnose_pdd_sample()}} combines the algorithms with data to diagnose
single patients.
}
