% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_criteria.R
\name{specify_criteria}
\alias{specify_criteria}
\title{Specify criteria operationalisations for PDD diagnosis}
\usage{
specify_criteria()
}
\value{
A tibble where each row represents one unique operationalisation
of PDD diagnostic criteria. Columns indicate the type of criteria, individual
test indexes, thresholds, and other relevant parameters.
}
\description{
Computes a Cartesian product of possible diagnostic components to define
all combinations of operationalised criteria for probable Parkinson’s disease
dementia (PDD). This function is a helper and is typically called within
\code{diagnose_pdd_sample}.
}
