% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_kappa.r
\name{summarise_kappa}
\alias{summarise_kappa}
\title{Prepare a list of algorithms stratified by IADL mesure}
\usage{
summarise_kappa(algorithms, concordance)
}
\arguments{
\item{algorithms}{A tibble containing raw summaries of PDD rates as
generated by \code{summarise_rates}.}

\item{concordance}{description}
}
\value{
A list with two components:
\describe{
\item{\code{faq_tot}}{A numeric vector containing all Cohen's kappas
of algorithm pairs with IADL deficit defined as FAQ total > 7.}
\item{\code{faq_9}}{A numeric vector containing all Cohen's kappas
of algorithm pairs with IADL deficit defined as FAQ item 9 > 1.}
\item{\code{cross}}{A numeric vector containing all Cohen's kappas
of algorithm pairs with different IADL deficit definitions.}
\item{\code{iadl}}{A numeric vector containing all Cohen's kappas
of algorithm pairs which differ only in IADL deficit definition.}
\item{\code{sums}}{A named character vector containing mean and
standard deviation of each vector of Kappas contained in this list.}
}
}
\description{
Using the table with PDD rates estimates per diagnostic algorithm,
this function extracts a list of algorithms using different IADL
criteria for PDD. If there was an algorithm that uses both criteria
at the same time, the function prints an error.
}
