% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_rates.R
\name{summarise_rates}
\alias{summarise_rates}
\title{Orders and prints algorithms based on their estimated PDD rates}
\usage{
summarise_rates(d0, vars, descending = TRUE)
}
\arguments{
\item{d0}{A list with PDD data generated via \code{diagnose_pdd_sample}.}

\item{vars}{A data.frame, tibble, or matrix with variable names in the first column,
variable labels in the second column, type of variable (continuous, binary, or nominal)
in the third column, and optionally a fourth column denoting group and a fifth column
mapping each label to its description in the table’s note.
Alternatively, a path to a CSV file (semicolon-delimited) containing such a table.}

\item{descending}{A logical indicating whether operationalisations with the highest
estimated PDD rates should be listed first (default is \code{TRUE}).}
}
\value{
A list with:
\describe{
\item{\code{table}}{A tibble containing raw summaries of PDD rates.}
\item{\code{plot}}{A \code{ggplot2} object visualising estimated PDD rates.}
\item{\code{gtable}}{A named list of gt tables:
\describe{
\item{\code{gtab_rates}{An APA-style \code{gt} table summarising the results.}}
\item{\code{gtab_algos}{An APA-style \code{gt} table listing algorithms used.}}
}}
}
}
\description{
This function computes summaries of the estimated PDD rates,
stratified by diagnostic operationalisation.
}
\examples{
\dontrun{
p <- data_paths("data-raw")
data <- prepare_data(p)
pdd  <- diagnose_pdd_sample(data)
vars <- here::here("data-raw", "VariablesOfInterest.csv")

rates0 <- summarise_rates(pdd, vars)
rates1 <- summarise_rates(pdd, vars, TRUE)  # rates0 and rates1 are identical
rates2 <- summarise_rates(pdd, vars, FALSE) # ascending order
}

}
